+++ ChangeLog.markdown
@@ -5,6 +5,7 @@
 * Improved filter values alpha numeric sorting, sort_as: "alphaNum"
 * Fixed autocomplete in cumulative mode filtering
 * Fixed chosen in cumulative mode filtering https://github.com/vedmack/yadcf/issues/344
+* Support jQuery 3 (for .dataTable({}).yadcf api) https://github.com/vedmack/yadcf/issues/360
 * Some of additional fixed issues https://github.com/vedmack/yadcf/issues/346